Revert "Enhancement: Account Center Updates (#1723)"
This reverts commit 0297ebb0298d59799852d51d9089cfcc09f37296.
