+++ docs/src/routes/docs/[...3]modules/[...1]core/+page.md
@@ -17,22 +17,6 @@ title: Core
 
 This is the core package that contains all of the UI and logic to be able to seamlessly connect user's wallets to your app and track the state of those wallets. Onboard no longer contains any wallet specific code, so wallets need to be passed in upon initialization.
 
-:::admonition type="tip"
-_note: Release 2.23.0 moves the default position of the account center from topRight to bottomRight. To reset your application to topRight, include the following when initializing onboard:_
-```typescript
-  accountCenter: {
-      desktop: {
-        enabled: true,
-        position: 'topRight'
-      },
-      mobile: {
-        enabled: true,
-        position: 'topRight'
-      }
-    }
-```
-:::
-
 ## Install
 
 Install the core module:
