+++ packages/core/README.md
@@ -6,20 +6,6 @@
 
 This is the core package that contains all of the UI and logic to be able to seamlessly connect user's wallets to your app and track the state of those wallets. Onboard no longer contains any wallet specific code, so wallets need to be passed in upon initialization.
 
-_Tip: Release 2.23.0 moves the default position of the account center from topRight to bottomRight. To reset your application to topRight, include the following when initializing onboard:_
-```typescript
-  accountCenter: {
-      desktop: {
-        enabled: true,
-        position: 'topRight'
-      },
-      mobile: {
-        enabled: true,
-        position: 'topRight'
-      }
-    }
-```
-
 ## Quick start
 
 Checkout our full library of quick start examples for connecting and interacting with EVM based wallets
