+++ packages/core/src/views/Index.svelte
@@ -378,6 +378,7 @@
   .container {
     padding: 16px;
     font-family: var(--onboard-font-family-normal, var(--font-family-normal));
+    width: 100%;
     pointer-events: none;
     touch-action: none;
   }
