+++ packages/core/src/views/account-center/Index.svelte
@@ -1,11 +1,14 @@
 <script lang="ts">
   import { onDestroy } from 'svelte'
   import { updateAccountCenter } from '../../store/actions.js'
+  import { wallets$ } from '../../streams.js'
   import { state } from '../../store/index.js'
   import { shareReplay, startWith } from 'rxjs/operators'
-  import AccountCenterPanel from './AccountCenterPanel.svelte'
-  import TriggerLarge from './AcctCenterTriggerLarge.svelte'
-  import TriggerSmall from './AcctCenterTriggerSmall.svelte'
+  import Maximized from './Maximized.svelte'
+  import Minimized from './Minimized.svelte'
+  import Micro from './Micro.svelte'
+
+  export let mountInContainer = false
 
   const accountCenter$ = state
     .select('accountCenter')
@@ -22,20 +25,19 @@
 
 <svelte:window on:click={minimize} />
 
-{#key $accountCenter$.position}
-{#key $accountCenter$.minimal}
-<div class="container flex flex-column items-end" style="width: 315px;">
-  {#if $accountCenter$.position.includes('bottom')}
-    <AccountCenterPanel />
+{#if mountInContainer}
+  {#if $wallets$.length}
+    <div class="container flex flex-column fixed z-indexed">
+      <svelte:self />
+    </div>
   {/if}
-  {#if $accountCenter$.minimal}
-    <TriggerSmall />
-  {:else}
-    <TriggerLarge />
-  {/if}
-  {#if $accountCenter$.position.includes('top')}
-    <AccountCenterPanel />
-  {/if}
-</div>
-{/key}
-{/key}
+{:else if !$accountCenter$.expanded && !$accountCenter$.minimal}
+  <!-- minimized -->
+  <Minimized />
+{:else if !$accountCenter$.expanded && $accountCenter$.minimal}
+  <!-- micro -->
+  <Micro />
+{:else}
+  <!-- maximized -->
+  <Maximized />
+{/if}
