+++ packages/core/src/views/account-center/Micro.svelte
@@ -1,46 +1,29 @@
 <script lang="ts">
-  import { shareReplay, startWith } from 'rxjs/operators'
   import { wallets$ } from '../../streams.js'
   import { updateAccountCenter } from '../../store/actions.js'
   import { questionIcon } from '../../icons/index.js'
   import { WalletAppBadge, SuccessStatusIcon } from '../shared/index.js'
-  import { state } from '../../store/index.js'
   import { configuration } from '../../configuration.js'
 
   const { appMetadata } = configuration
   const appIcon = (appMetadata && appMetadata.icon) || questionIcon
-
   $: [primaryWallet] = $wallets$
 
-  const accountCenter$ = state
-    .select('accountCenter')
-    .pipe(startWith(state.get().accountCenter), shareReplay(1))
-
-  function toggle() {
-    updateAccountCenter({ expanded: !$accountCenter$.expanded })
+  function maximize() {
+    updateAccountCenter({ expanded: true })
   }
 </script>
 
 <style>
   .minimized {
-    --background-color: var(
-      --account-center-minimized-background,
-      var(--w3o-background-color, white)
-    );
-    --text-color: var(--w3o-text-color, var(--gray-700));
-    --border-color: var(
-      --account-center-border,
-      var(--w3o-border-color, var(--onboard-gray-200, var(--gray-200)))
-    );
-    --border-radius: var(
-      --account-center-border-radius,
-      var(--w3o-border-radius, 1rem)
-    );
+    --background-color: var(--account-center-micro-background, var(--w3o-background-color, white));
+    --text-color: var(--w3o-text-color);
+    --border-color: var(--account-center-border, var(--w3o-border-color, var(--onboard-gray-200, var(--gray-200))));
+    --border-radius: var(--account-center-border-radius, var(--w3o-border-radius, 1rem));
 
     cursor: pointer;
     pointer-events: auto;
-    width: 80px;
-    border: 1px solid;
+    border: 1px solid transparent;
 
     background: var(--background-color);
     color: var(--text-color);
@@ -50,7 +33,17 @@
       --account-center-box-shadow,
       var(--onboard-shadow-3, var(--shadow-3))
     );
-    z-index: 1;
+  }
+
+  .drop-shadow {
+    filter: drop-shadow(0px 1px 4px rgba(0, 0, 0, 0.2));
+  }
+
+  .inner-row {
+    display: flex;
+    flex-flow: row nowrap;
+    width: 80px;
+    padding: 0.75rem;
   }
 
   .wallet-square-wrapper {
@@ -63,25 +56,9 @@
     right: -4px;
     bottom: -4px;
   }
-
-  .inner-row {
-    display: flex;
-    flex-flow: row nowrap;
-    width: 80px;
-    padding: 0.75rem;
-  }
-  .drop-shadow {
-    filter: drop-shadow(0px 1px 4px rgba(0, 0, 0, 0.2));
-  }
 </style>
 
-<div
-  class="minimized"
-  on:click|stopPropagation={toggle}
-  style={$accountCenter$.position.includes('Left')
-    ? 'align-self: flex-start'
-    : null}
->
+<div class="minimized" on:click|stopPropagation={maximize}>
   <div class="inner-row">
     <!-- app and wallet icon badge -->
     <div class="drop-shadow">
