+++ packages/core/src/views/account-center/Minimized.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-  import { shareReplay, startWith } from 'rxjs/operators'
   import { fade } from 'svelte/transition'
   import { wallets$ } from '../../streams.js'
   import {
@@ -55,12 +54,8 @@
 
   $: defaultChainStyles = getDefaultChainStyles(primaryChain && primaryChain.id)
 
-  const accountCenter$ = state
-    .select('accountCenter')
-    .pipe(startWith(state.get().accountCenter), shareReplay(1))
-
-  function toggle() {
-    updateAccountCenter({ expanded: !$accountCenter$.expanded })
+  function maximize() {
+    updateAccountCenter({ expanded: true })
   }
 </script>
 
@@ -94,7 +89,6 @@
       --account-center-box-shadow,
       var(--onboard-shadow-3, var(--shadow-3))
     );
-    z-index: 1;
   }
 
   .inner-row {
@@ -155,7 +149,7 @@
   in:fade={{ duration: 250 }}
   out:fade={{ duration: 100 }}
   class="minimized"
-  on:click|stopPropagation={toggle}
+  on:click|stopPropagation={maximize}
 >
   <div class="inner-row">
     <div class="flex relative">
